removed floating point operations form algorithm

+++ python-wrapper/sensirion_gas_index_algorithm/gas_index_algorithm.py
@@ -13,7 +13,7 @@ log = logging.getLogger(__name__)
 class GasIndexAlgorithm(object):
 
     def __init__(self, algorithm_type: int,
-                 sampling_interval: float = sensirion_gas_index_algorithm_wrap.default_sampling_interval):
+                 sampling_interval: int = sensirion_gas_index_algorithm_wrap.default_sampling_interval):
         self._algorithm_type = algorithm_type
         self._algorithm = sensirion_gas_index_algorithm_wrap.GasIndexAlgorithmParams(algorithm_type, sampling_interval)
 
@@ -21,7 +21,7 @@ class GasIndexAlgorithm(object):
         """Get the version of the engine."""
         return self._algorithm.get_version()
 
-    def init(self, sampling_interval: float = sensirion_gas_index_algorithm_wrap.default_sampling_interval):
+    def init(self, sampling_interval: int = sensirion_gas_index_algorithm_wrap.default_sampling_interval):
         """Re-initialize / reset the engine."""
         self._algorithm.init(self._algorithm_type, sampling_interval)
